use explicit typecase in coulomb table lookup in a few more cases

+++ src/DIELECTRIC/pair_lj_cut_coul_long_dielectric.cpp
@@ -271,7 +271,7 @@ double PairLJCutCoulLongDielectric::single(int i, int j, int itype, int jtype, d
       rsq_lookup_single.f = rsq;
       itable = rsq_lookup_single.i & ncoulmask;
       itable >>= ncoulshiftbits;
-      fraction = (rsq_lookup_single.f - rtable[itable]) * drtable[itable];
+      fraction = ((double) rsq_lookup_single.f - rtable[itable]) * drtable[itable];
       table = ftable[itable] + fraction * dftable[itable];
       forcecoul = q[i] * q[j] * table;
       if (factor_coul < 1.0) {
